Refactor and tweaks for semantics
