+++ src/Router/router.tsx
@@ -4,12 +4,12 @@ import {
     Route,
 } from 'react-router-dom';
 
-import Base from 'src/Layout/Base';
+import AppLayout from 'src/Layout/AppLayout';
 import Transactions from 'src/Pages/Transactions';
 
 const router = createBrowserRouter(
     createRoutesFromElements(
-        <Route element={<Base />}>
+        <Route element={<AppLayout />}>
             <Route path="/" element={<Transactions />} />
         </Route>
     )
